% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flags.R
\name{Finland}
\alias{Finland}
\title{R Flags}
\usage{
Finland(a = 20000)
}
\arguments{
\item{n}{Number of points}
}
\value{
Country flag with a fixed number of points.
}
\description{
R flag is a simple drawing game. 
Players can choose the number of points and name of a country to get a country flag.
}
\note{
Welcome to add more flags
}
\author{
Wei Wang
}
