% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{little_game}
\alias{little_game}
\title{The profile of this package}
\usage{
little_game(name)
}
\arguments{
\item{name}{The name of a game.("mine_sweeper", "gomoku", "r_snake", "r_flags")}
}
\value{
A list of information about the game.
}
\description{
This function contains some information regarding the four games in this 
           package. If you have doubt about these games, this function will probably help you.
}
\examples{
little_game("Snake")
}
