% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mine_sweeper_pkg.R
\name{minesweeper}
\alias{minesweeper}
\title{Play the Mine Sweeper game in R
Just type in minesweeper() and the game will start
The controls should be familiar to you: Click the left mouse button to dig in
an area, and right button to mark or unmark the area with flags.}
\usage{
minesweeper()
}
\arguments{
\item{width}{number of grids in horizontal axis}

\item{height}{number of grids in vertical axis}

\item{mines}{number of mines}

\item{cheat}{logical. If \code{TRUE} a matrix indicating the mines will be
printed}
}
\description{
Play the Mine Sweeper game in R
Just type in minesweeper() and the game will start
The controls should be familiar to you: Click the left mouse button to dig in
an area, and right button to mark or unmark the area with flags.
}
\note{
Linux/Mac users have to use \code{X11(type = 'Xlib')} or the Cairo
  graphics device \code{Cairo()} in the package \pkg{cairoDevice}.
}
\references{
\url{http://en.wikipedia.org/wiki/Minesweeper_(computer_game)}
}
\author{
Bowei Wei \email{wbwlazycat@126.com}
}
