% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snake.R
\name{stage1}
\alias{stage1}
\title{R Snake}
\usage{
stage1()
}
\arguments{
\item{Direction}{operator}

\item{Direction}{operator}
}
\value{
stage1
R Snake

furit
R Snake

Check the failure
R Snake

Snake head
R Snake

Snake body
R Snake

The canvas background
R Snake

matrix with data
}
\description{
R Snake
}
